requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
scalar-fastapi>=1.4.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pgvector>=0.2.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM Integration
llama-cpp-python>=0.2.0